Moved currency exchange for stock prices into a separate class, cleaned code, added gitignore file to the project

+++ Source Code/app/src/main/java/redditandroidapp/data/database/CompanyDatabaseEntity.kt
@@ -13,7 +13,7 @@ data class CompanyDatabaseEntity(
         val previousQuarter_NetIncome: Double?,
         val recentQuarter_GrossProfit: Double?,
         val recentQuarter_NetIncome: Double?,
-        val eps: Float?,
+        val eps: Double?,
 
         val today_OutstandingShares: Long?,
         val today_SharePrice: Double?
@@ -55,7 +55,7 @@ data class CompanyDatabaseEntity(
         }
     }
 
-    fun getEarningsPerShare(): Float? {
+    fun getEarningsPerShare(): Double? {
         return eps
     }
 }
